Moving Screens and added about libraries screen
