+++ MultiPlatformLibrary/src/commonMain/kotlin/org/rhasspy/mobile/nativeutils/LibrariesUtils.kt
@@ -1,0 +1,5 @@
+package org.rhasspy.mobile.nativeutils
+
+import com.mikepenz.aboutlibraries.entity.Library
+
+expect suspend fun loadLibraries(): List<Library>
