+++ MultiPlatformLibrary/src/iosMain/kotlin/org/rhasspy/mobile/nativeutils/LibrariesUtils.kt
@@ -1,0 +1,7 @@
+package org.rhasspy.mobile.nativeutils
+
+import com.mikepenz.aboutlibraries.entity.Library
+
+actual suspend fun loadLibraries(): List<Library> {
+    TODO("Not yet implemented")
+}
