+++ androidApp/src/main/java/org/rhasspy/mobile/android/permissions/MicrophonePermission.kt
@@ -1,21 +1,32 @@
 package org.rhasspy.mobile.android.permissions
 
 import androidx.activity.ComponentActivity
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.core.tween
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.foundation.background
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Mic
+import androidx.compose.material.icons.filled.MicOff
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.unit.dp
 import dev.icerock.moko.resources.StringResource
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import org.rhasspy.mobile.MR
 import org.rhasspy.mobile.android.utils.Text
+import org.rhasspy.mobile.android.utils.observe
 import org.rhasspy.mobile.android.utils.translate
 import org.rhasspy.mobile.nativeutils.MicrophonePermission
+import org.rhasspy.mobile.viewModels.HomeScreenViewModel
 
 /**
  * 3 parts where this is shown:
@@ -24,6 +35,31 @@ import org.rhasspy.mobile.nativeutils.MicrophonePermission
  * local WakeWord service
  * click on check audio level button
  */
+@Composable
+fun MicrophonePermissionRequired(viewModel: HomeScreenViewModel, snackbarHostState: SnackbarHostState) {
+    AnimatedVisibility(
+        enter = fadeIn(animationSpec = tween(50)),
+        exit = fadeOut(animationSpec = tween(50)),
+        visible = viewModel.isMicrophonePermissionRequestRequired.observe()
+    ) {
+        val microphonePermission = requestMicrophonePermission(snackbarHostState, MR.strings.microphonePermissionInfoWakeWord) {}
+
+        IconButton(
+            onClick = { microphonePermission.invoke() },
+            modifier = Modifier.background(
+                color = MaterialTheme.colorScheme.errorContainer,
+                shape = RoundedCornerShape(8.dp)
+            )
+        )
+        {
+            org.rhasspy.mobile.android.utils.Icon(
+                imageVector = Icons.Filled.MicOff,
+                tint = MaterialTheme.colorScheme.onErrorContainer,
+                contentDescription = MR.strings.microphone
+            )
+        }
+    }
+}
 
 @Composable
 fun requestMicrophonePermission(
