+++ buildSrc/src/main/kotlin/Mikepenz.kt
@@ -1,0 +1,8 @@
+import de.fayard.refreshVersions.core.DependencyGroup
+
+object Mikepenz : DependencyGroup(group = "com.mikepenz") {
+
+    val aboutLibrariesCore = module("aboutlibraries-core")
+    val aboutLibrariesCompose = module("aboutlibraries-compose")
+
+}
