+++ versions.properties
@@ -6,6 +6,7 @@
 ####
 #### suppress inspection "SpellCheckingInspection" for whole file
 #### suppress inspection "UnusedProperty" for whole file
+
 plugin.android=7.1.3
 ## # available=7.2.0-alpha01
 ## # available=7.2.0-alpha02
@@ -27,7 +28,9 @@ plugin.android=7.1.3
 ## # available=7.3.0-alpha06
 ## # available=7.3.0-alpha07
 ## # available=7.3.0-alpha08
+
 version.ai.picovoice..porcupine-android=2.1.0
+
 version.androidx.activity=1.4.0
 ##            # available=1.5.0-alpha01
 ##            # available=1.5.0-alpha02
@@ -36,14 +39,23 @@ version.androidx.activity=1.4.0
 ##            # available=1.5.0-alpha05
 ##            # available=1.5.0-beta01
 ##            # available=1.6.0-alpha01
+
 version.androidx.compose.compiler=1.2.0-alpha08
+
 version.androidx.compose.foundation=1.2.0-alpha08
+
 version.androidx.compose.material=1.2.0-alpha08
+
 version.androidx.compose.material3=1.0.0-alpha10
+
 version.androidx.compose.runtime=1.2.0-alpha08
+
 version.androidx.compose.ui=1.2.0-alpha08
+
 version.androidx.constraintlayout-compose=1.0.0
+
 version.androidx.core-splashscreen=1.0.0-beta02
+
 version.androidx.lifecycle=2.4.1
 ##             # available=2.5.0-alpha01
 ##             # available=2.5.0-alpha02
@@ -52,6 +64,7 @@ version.androidx.lifecycle=2.4.1
 ##             # available=2.5.0-alpha05
 ##             # available=2.5.0-alpha06
 ##             # available=2.5.0-beta01
+
 version.androidx.lifecycle-viewmodel-compose=2.4.1
 ##                               # available=2.5.0-alpha01
 ##                               # available=2.5.0-alpha02
@@ -60,30 +73,55 @@ version.androidx.lifecycle-viewmodel-compose=2.4.1
 ##                               # available=2.5.0-alpha05
 ##                               # available=2.5.0-alpha06
 ##                               # available=2.5.0-beta01
+
 version.androidx.multidex=2.0.1
+
 version.androidx.navigation-compose=2.4.2
 ##                      # available=2.5.0-alpha01
 ##                      # available=2.5.0-alpha02
 ##                      # available=2.5.0-alpha03
 ##                      # available=2.5.0-alpha04
 ##                      # available=2.5.0-beta01
+
 ## unused
 version.androidx.savedstate=1.1.0
 ##              # available=1.2.0-alpha01
 ##              # available=1.2.0-alpha02
 ##              # available=1.2.0-beta01
+
 version.androidx.window=1.0.0
 ##          # available=1.1.0-alpha01
+
 version.br.com.devsrsouza.compose.icons.android..font-awesome=1.0.0
+
 version.com.benasher44..uuid=0.4.0
+
+version.com.mikepenz..aboutlibraries=10.1.0
+
+version.com.mikepenz..aboutlibraries-compose=10.1.0
+
+version.com.mikepenz..aboutlibraries-core=10.1.0
+
+version.com.mikepenz..aboutlibraries-core-android-debug=10.1.0
+
+version.com.mikepenz.aboutlibraries.plugin..aboutlibraries-plugin=10.1.0
+
 version.dev.icerock.moko..mvvm-core=0.12.0
+
 version.dev.icerock.moko..mvvm-livedata=0.12.0
+
 version.dev.icerock.moko..mvvm-livedata-resources=0.12.0
+
 version.dev.icerock.moko..mvvm-state=0.12.0
+
 version.dev.icerock.moko..permissions=0.11.0
+
 version.dev.icerock.moko..resources=0.19.0
+
 version.dev.icerock.moko..resources-compose=0.19.0
+
 version.dev.icerock.moko..resources-generator=0.19.0
+
 version.google.accompanist=0.23.1
 ##             # available=0.24.0-alpha
 ##             # available=0.24.1-alpha
@@ -92,6 +130,7 @@ version.google.accompanist=0.23.1
 ##             # available=0.24.4-alpha
 ##             # available=0.24.5-alpha
 ##             # available=0.24.6-alpha
+
 version.google.android.material=1.5.0
 ##                  # available=1.6.0-alpha01
 ##                  # available=1.6.0-alpha02
@@ -99,20 +138,29 @@ version.google.android.material=1.5.0
 ##                  # available=1.6.0-beta01
 ##                  # available=1.6.0-rc01
 ##                  # available=1.7.0-alpha01
+
 version.kermit=1.1.1
+
 version.kotlin=1.6.20
 ## # available=1.6.21
+
 version.kotlinx.coroutines=1.6.1
+
 version.kotlinx.datetime=0.3.2
+
 version.kotlinx.serialization=1.3.2
+
 ### 2.0.0 server not working on android
-version.ktor=2.0.0
+ version.ktor=2.0.0
 ### available=2.0.0-beta-1
 ### available=2.0.0-rc-1
 ### available=2.0.0
+
 ## unused
 version.org.jacoco..org.jacoco.ant=0.8.3
+
 version.multiplatform-settings=0.8.1
+
 version.org.slf4j..slf4j-simple=1.7.36
 ##                  # available=1.8.0-alpha0
 ##                  # available=1.8.0-alpha1
